[optimization] Move analysis package into optimization
