--- samlang-core-analysis/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "samlang-core-analysis",
-  "version": "0.0.1",
-  "license": "AGPLv3",
-  "scripts": {
-    "compile": "tsc --incremental --noEmit"
-  },
-  "dependencies": {
-    "samlang-core-ast": "workspace:0.0.1",
-    "samlang-core-utils": "workspace:0.0.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.20",
-    "typescript": "^4.1.3"
-  }
-}
