--- samlang-core-analysis/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../configurations/tsconfig.base.json"
-}
