+++ samlang-core-optimization/__tests__/control-flow-graph.test.ts
