+++ samlang-core-optimization/package.json
@@ -6,7 +6,6 @@
     "compile": "tsc --incremental --noEmit"
   },
   "dependencies": {
-    "samlang-core-analysis": "workspace:0.0.1",
     "samlang-core-ast": "workspace:0.0.1",
     "samlang-core-utils": "workspace:0.0.1"
   },
