--- workspaces.json
@@ -1,208 +1,0 @@
-{
-  "__type__": "@generated",
-  "information": {
-    "@dev-sam/samlang-cli": {
-      "workspaceLocation": "samlang-cli",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-printer",
-        "samlang-core-checker",
-        "samlang-core-services",
-        "@dev-sam/samlang-cli"
-      ]
-    },
-    "@dev-sam/samlang-demo": {
-      "workspaceLocation": "samlang-demo",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-interpreter",
-        "samlang-core-printer",
-        "samlang-core-checker",
-        "samlang-core-services",
-        "@dev-sam/samlang-demo"
-      ]
-    },
-    "samlang-core-analysis": {
-      "workspaceLocation": "samlang-core-analysis",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis"
-      ]
-    },
-    "samlang-core-ast": {
-      "workspaceLocation": "samlang-core-ast",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast"
-      ]
-    },
-    "samlang-core-checker": {
-      "workspaceLocation": "samlang-core-checker",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-checker"
-      ]
-    },
-    "samlang-core-compiler": {
-      "workspaceLocation": "samlang-core-compiler",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler"
-      ]
-    },
-    "samlang-core-errors": {
-      "workspaceLocation": "samlang-core-errors",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors"
-      ]
-    },
-    "samlang-core-integration-tests": {
-      "workspaceLocation": "samlang-core-integration-tests",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-interpreter",
-        "samlang-core-printer",
-        "samlang-core-checker",
-        "samlang-core-services",
-        "samlang-core-integration-tests"
-      ]
-    },
-    "samlang-core-interpreter": {
-      "workspaceLocation": "samlang-core-interpreter",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-interpreter"
-      ]
-    },
-    "samlang-core-optimization": {
-      "workspaceLocation": "samlang-core-optimization",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization"
-      ]
-    },
-    "samlang-core-parser": {
-      "workspaceLocation": "samlang-core-parser",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser"
-      ]
-    },
-    "samlang-core-parser-generated": {
-      "workspaceLocation": "samlang-core-parser-generated",
-      "dependencyChain": [
-        "samlang-core-parser-generated"
-      ]
-    },
-    "samlang-core-printer": {
-      "workspaceLocation": "samlang-core-printer",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-printer"
-      ]
-    },
-    "samlang-core-services": {
-      "workspaceLocation": "samlang-core-services",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-checker",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler",
-        "samlang-core-services"
-      ]
-    },
-    "samlang-core-utils": {
-      "workspaceLocation": "samlang-core-utils",
-      "dependencyChain": [
-        "samlang-core-utils"
-      ]
-    },
-    "samlang-dummy": {
-      "workspaceLocation": "samlang-dummy",
-      "dependencyChain": [
-        "samlang-core-utils",
-        "samlang-core-ast",
-        "samlang-core-analysis",
-        "samlang-core-optimization",
-        "samlang-core-compiler",
-        "samlang-core-errors",
-        "samlang-core-parser-generated",
-        "samlang-core-parser",
-        "samlang-core-printer",
-        "samlang-core-checker",
-        "samlang-core-services",
-        "@dev-sam/samlang-cli",
-        "samlang-core-interpreter",
-        "@dev-sam/samlang-demo",
-        "samlang-dummy"
-      ]
-    }
-  },
-  "topologicallyOrdered": [
-    "samlang-core-utils",
-    "samlang-core-ast",
-    "samlang-core-analysis",
-    "samlang-core-optimization",
-    "samlang-core-compiler",
-    "samlang-core-errors",
-    "samlang-core-parser-generated",
-    "samlang-core-parser",
-    "samlang-core-printer",
-    "samlang-core-checker",
-    "samlang-core-services",
-    "@dev-sam/samlang-cli",
-    "samlang-core-interpreter",
-    "samlang-core-integration-tests",
-    "@dev-sam/samlang-demo",
-    "samlang-dummy"
-  ]
-}
